refactor: re-org, lua namespace dir, after
Re-organize things, namespace overrides & main lua configuration under
named directory (vin).  Also, utilize the after dir for plugin setup &
plugin mapping overrides -- this has a nice sideffect where, if I bork
some config, the other plugins still load fine
