+++ after/plugin/alpha.lua
@@ -1,4 +1,3 @@
--- =============== DASHBOARD ===================
 local alpha = require('alpha')
 local dashboard = require('alpha.themes.dashboard')
 
@@ -39,116 +38,3 @@ dashboard.section.buttons.val = {
 }
 
 alpha.setup(dashboard.opts)
-
--- =============== STATUSLINE ==================
-require('lualine').setup {
-  options = {
-    theme = 'tokyonight';
-    icons_enabled = true;
-    section_separators = {'', ''};
-    component_separators = {'|', '|'};
-  }
-}
-
--- =============== GUIDES =======================
-
--- Config: indent-blankline
-vim.g.indent_blankline_use_treesitter = true
-vim.g.indent_blankline_filetype = {
-  'lua',
-  'javascript',
-  'javascriptreact',
-  'typescript',
-  'typescriptreact',
-  'html',
-  'nix',
-  'rust',
-  'svelte',
-  'css',
-  'scss',
-  'c',
-  'python',
-}
-
-require('indent_blankline').setup {
-  space_char_blankline = " ",
-  show_current_context = true,
-  show_current_context_start = true,
-}
-
--- =============== COLORSCHEMES ================
-
--- Config: moonfly
--- vim.g.moonflyTransparent = 1
-
--- Config: catppuccin
--- local catppuccin = require('catppuccin')
-
--- catppuccin.setup({
---   flavour = 'mocha',
---   transparent_background = false,
---   -- integrations = {
---   --   telescope = true,
---   --   gitsigns = true,
---   --   cmp = true,
---   --   treesitter = true,
---   -- },
---   highlight = {
---     enable = true,
---     additional_vim_regex_highlightling = false,
---   },
---   term_colors = true,
---   dim_inactive = {
---     enabled = true,
---     shade = 'dark',
---     percentage = 0.15,
---   },
---   native_lsp = {
---     enabled = true,
---     virtual_text = {
---         errors = { "italic" },
---         hints = { "italic" },
---         warnings = { "italic" },
---         information = { "italic" },
---     },
---     underlines = {
---         errors = { "underline" },
---         hints = { "underline" },
---         warnings = { "underline" },
---         information = { "underline" },
---     },
---   },
--- })
-
--- vim.api.nvim_command "colorscheme catppuccin"
-
--- Config tokyonight
-require("tokyonight").setup({
-  style = "night",
-  transparent = true,
-})
-
-vim.cmd[[ colorscheme tokyonight ]]
-
--- =============== TROUBLE ================
-require('todo-comments').setup({
-  keywords = {
-    FIX = {
-      color = 'error', -- can be a hex color, or a named color (see below)
-      alt = { 'FIXME', 'BUG', 'FIXIT', 'ISSUE', 'fix', 'fixme', 'bug' }, -- a set of other keywords that all map to this FIX keywords
-      -- signs = false, -- configure signs for some keywords individually
-    },
-    TODO = { color = 'info' },
-    HACK = { color = 'warning' },
-    WARN = { color = 'warning', alt = { 'WARNING', 'XXX' } },
-    PERF = { alt = { 'OPTIM', 'PERFORMANCE', 'OPTIMIZE' } },
-    NOTE = { color = 'hint', alt = { 'INFO' } },
-  },
-  colors = {
-    error = { 'DiagnosticError', 'ErrorMsg', '#DC2626' },
-    warning = { 'DiagnosticWarning', 'WarningMsg', '#FBBF24' },
-    info = { 'DiagnosticInformation', '#2563EB' },
-    hint = { 'DiagnosticHint', '#10B981' },
-    default = { 'Identifier', '#7C3AED' },
-  },
-})
