+++ after/plugin/lir.lua
@@ -1,0 +1,62 @@
+local actions = require'lir.actions'
+local mark_actions = require 'lir.mark.actions'
+local clipboard_actions = require'lir.clipboard.actions'
+
+local map = require('vin.utils').map
+
+-- Config: lir
+require'lir'.setup {
+  show_hidden_files = true,
+  devicons_enable = true,
+  mappings = {
+    ['l']     = actions.edit,
+    ['<CR>']  = actions.edit,
+    ['<C-s>'] = actions.split,
+    ['<C-v>'] = actions.vsplit,
+    ['<C-t>'] = actions.tabedit,
+
+    ['h']     = actions.up,
+    ['q']     = actions.quit,
+
+    ['K']     = actions.mkdir,
+    ['N']     = actions.newfile,
+    ['R']     = actions.rename,
+    ['@']     = actions.cd,
+    ['Y']     = actions.yank_path,
+    ['.']     = actions.toggle_show_hidden,
+    ['D']     = actions.delete,
+
+    ['J'] = function()
+      mark_actions.toggle_mark()
+      vim.cmd('normal! j')
+    end,
+    ['C'] = clipboard_actions.copy,
+    ['X'] = clipboard_actions.cut,
+    ['P'] = clipboard_actions.paste,
+  },
+  float = {
+    winblend = 0,
+    win_opts = function()
+      local width = math.floor(vim.o.columns * 0.8)
+      local height = math.floor(vim.o.lines * 0.8)
+
+      return {
+        border = require("lir.float.helper").make_border_opts({
+          "+", "─", "+", "│", "+", "─", "+", "│",
+        }, "Normal"),
+        width = width,
+        height = height,
+        row = 1,
+        col = math.floor((vim.o.columns - width) / 2),
+      }
+    end
+  },
+  hide_cursor = false,
+}
+
+-- Config: lir-git-status
+require'lir.git_status'.setup({ show_ignored = false })
+
+-- === Mappings ===
+map('n', '<Leader>F', ":lua require'lir.float'.toggle()<CR>")
+map('n', '-', [[<CMD>execute 'e ' .. expand('%:p:h')<CR>]]) -- netrw/dirvish file exp functionality
