+++ after/plugin/lualine.lua
@@ -1,0 +1,8 @@
+require('lualine').setup {
+  options = {
+    theme = 'tokyonight';
+    icons_enabled = true;
+    section_separators = {'', ''};
+    component_separators = {'|', '|'};
+  }
+}
