+++ lua/vin/init.lua
@@ -1,3 +1,9 @@
+require('vin.set')
+require('vin.remap')
+require('vin.commands')
+require('vin.packer')
+
+-- Disable Default BuiltIn Plugins
 local disabled_built_ins = {
   'netrw',
   'netrwPlugin',
